feat(bit): 添加DoubleBitView的规格化数判断功能并修复位操作类型问题
添加Normalized()方法用于判断双精度浮点数是否为规格化数
修复Bit()函数中的位操作类型转换问题
添加相关头文件引用和测试代码

+++ include/base/bit/bit.h
@@ -112,7 +112,7 @@ namespace base
 		constexpr uint64_t Bit(int bit_index)
 		{
 			CheckBitIndex<uint64_t>(bit_index);
-			return 0b1 << bit_index;
+			return static_cast<uint64_t>(0b1) << bit_index;
 		}
 
 		///
